+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Percent succeeded proposals
+
 ## [0.0.34] - 2023-10-20
 
 ### Added